Moved raw code sections into their own file, including them with `include_str!`.
