+++ src/preface_former.rs
@@ -1,0 +1,4 @@
+#![allow(unused_imports)]
+#![allow(dead_code)]
+
+use std::io::{stdout, stdin, Read, Write};
